Reorganize project into a `lib` system.  Separate concerns for days and inputs.  Add YARD for documentation
