+++ lib/advent_of_code/days.rb
@@ -1,0 +1,5 @@
+module AdventOfCode
+  module Days
+    Dir.glob(File.expand_path('days/**', __dir__)).each { |file| require "#{file}"}
+  end
+end
