+++ lib/advent_of_code/inputs.rb
@@ -1,0 +1,5 @@
+module AdventOfCode
+  module Inputs
+    Dir.glob(File.expand_path('inputs/**', __dir__)).each { |file| require "#{file}"}
+  end
+end
